tentando convex hull

+++ trem/1/e.cpp
@@ -24,6 +24,6 @@ const int INF = 0x3f3f3f3f;
 const ll LINF = 0x3f3f3f3f3f3f3f3fll;
 
 int main(){ _
-    int n, k; cin >> n >> k;    
+    ll n, k; cin >> n >> k;    
     exit(0);
 }
